# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "INCES LMS API",
    "version": "0.1.0",
    "description": "API del LMS del CFS Nacional de Soldadura \"Rafael Urdaneta\" (INCES, La Isabelica).\n\nEste documento se **genera** a partir de los esquemas Zod en `src/http/esquemas.ts`.\nNo se edita a mano: cualquier cambio manual se perderá en la siguiente generación.\nLos esquemas Zod son el único contrato; esto es una proyección suya.\n\nTodas las rutas exigen un token de Supabase Auth, salvo las sondas de salud.",
    "license": {
      "name": "Uso académico — TEG IUTEPI"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desarrollo local"
    },
    {
      "url": "https://api.example.com",
      "description": "Producción (pendiente de dominio)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token de acceso de Supabase Auth. El backend lo verifica contra GoTrue en cada petición."
      }
    },
    "schemas": {
...
      "CuerpoError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "codigo": {
                "type": "string"
              },
              "mensaje": {
                "type": "string"
              },
              "detalles": {}
            },
            "required": [
              "codigo",
              "mensaje"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "RespuestaModulos": {
        "type": "object",
        "properties": {
          "modulos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuloSistema"
            }
          }
        },
        "required": [
          "modulos"
        ]
      },
      "RespuestaModulo": {
        "type": "object",
        "properties": {
          "modulo": {
            "$ref": "#/components/schemas/ModuloSistema"
          }
        },
        "required": [
          "modulo"
        ]
      },
...
      "RespuestaParametros": {
        "type": "object",
        "properties": {
          "parametros": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParametroSistema"
            }
          }
        },
        "required": [
          "parametros"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/modulos": {
      "get": {
        "tags": [
          "Sesión"
        ],
        "summary": "Módulos visibles para el rol del llamante",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista filtrada por rol y por estado de habilitación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaModulos"
                }
              }
            }
          },
          "401": {
            "description": "Falta la sesión o el token no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "403": {
            "description": "La sesión es válida pero el rol no alcanza, o la cuenta está inactiva.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "503": {
            "description": "El servicio o la base de datos no están disponibles.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/modulos": {
      "get": {
        "tags": [
          "Administrador Maestro"
        ],
        "summary": "Todos los módulos, incluidos los apagados",
        "description": "A diferencia de `/api/v1/modulos`, no filtra por rol ni por habilitación: es la vista del panel.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Los nueve módulos del sistema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaModulos"
                }
              }
            }
          },
          "401": {
            "description": "Falta la sesión o el token no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "403": {
            "description": "La sesión es válida pero el rol no alcanza, o la cuenta está inactiva.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "503": {
            "description": "El servicio o la base de datos no están disponibles.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/modulos/{clave}": {
      "patch": {
        "tags": [
          "Administrador Maestro"
        ],
        "summary": "Enciende, apaga o reordena un módulo",
        "description": "Apagar `m0_cpanel` devuelve 409 MODULO_CRITICO: dejaría al administrador fuera del sistema. La base de datos tiene un trigger que impone la misma regla, para que no se pueda eludir desde el editor SQL.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Clave del módulo o del parámetro.",
              "example": "m0_cpanel"
            },
            "required": true,
            "name": "clave",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CuerpoCambiosModulo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Módulo actualizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaModulo"
                }
              }
            }
          },
          "400": {
            "description": "Petición inválida. El cuerpo o los parámetros no pasan la validación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "401": {
            "description": "Falta la sesión o el token no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "403": {
            "description": "La sesión es válida pero el rol no alcanza, o la cuenta está inactiva.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "404": {
            "description": "El recurso no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "409": {
            "description": "Se intentó apagar un módulo crítico.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "503": {
            "description": "El servicio o la base de datos no están disponibles.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/parametros": {
      "get": {
        "tags": [
          "Administrador Maestro"
        ],
        "summary": "Todos los parámetros, públicos y privados",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Listado completo de parámetros.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaParametros"
                }
              }
            }
          },
          "401": {
            "description": "Falta la sesión o el token no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "403": {
            "description": "La sesión es válida pero el rol no alcanza, o la cuenta está inactiva.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "503": {
            "description": "El servicio o la base de datos no están disponibles.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/parametros/{clave}": {
      "patch": {
        "tags": [
          "Administrador Maestro"
        ],
        "summary": "Cambia el valor de un parámetro",
        "description": "El valor debe encajar con el `tipo` declarado del parámetro. Sin esa comprobación, `max_faltas_consecutivas` podría quedar guardado como \"tres\" y el módulo de asistencia empezaría a comportarse de forma impredecible lejos del sitio donde se cometió el error.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Clave del módulo o del parámetro.",
              "example": "m0_cpanel"
            },
            "required": true,
            "name": "clave",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CuerpoValorParametro"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parámetro actualizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaParametro"
                }
              }
            }
          },
          "400": {
            "description": "Petición inválida. El cuerpo o los parámetros no pasan la validación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "401": {
            "description": "Falta la sesión o el token no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "403": {
            "description": "La sesión es válida pero el rol no alcanza, o la cuenta está inactiva.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "404": {
            "description": "El recurso no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "503": {
            "description": "El servicio o la base de datos no están disponibles.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auditoria": {
      "get": {
        "tags": [
          "Administrador Maestro"
        ],
        "summary": "Últimas entradas del registro de auditoría",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "description": "Máximo de entradas a devolver. Por defecto 50."
            },
            "required": false,
            "name": "limite",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Entradas más recientes primero.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaAuditoria"
                }
              }
            }
          },
          "400": {
            "description": "Petición inválida. El cuerpo o los parámetros no pasan la validación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "401": {
            "description": "Falta la sesión o el token no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "403": {
            "description": "La sesión es válida pero el rol no alcanza, o la cuenta está inactiva.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "503": {
            "description": "El servicio o la base de datos no están disponibles.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/acceso": {
      "get": {
        "tags": [
          "Administrador Maestro"
        ],
        "summary": "Traza de accesos (logins y fallos de autenticación) con filtros",
        "description": "Lee `auth_logs` para que el administrador audite inicios de sesión, IPs y fallos de autenticación. Soporta filtro por resultado, correo o usuario, y paginación. El filtro y el recorte se aplican en la base de datos; `total` es el número de filas que cumplen el filtro, no las devueltas.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "SUCCESS",
                "FAILED"
              ],
              "description": "Filtra por resultado del intento. Sin él, ambos."
            },
            "required": false,
            "name": "estado",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filtra por correo exacto (no parcial)."
            },
            "required": false,
            "name": "email",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Filtra por UUID de usuario de auth. Un valor no UUID se rechaza con 400."
            },
            "required": false,
            "name": "userId",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "description": "Tamaño de página. Por defecto 25."
            },
            "required": false,
            "name": "limite",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0,
              "description": "Filas a saltar antes de esta página. Por defecto 0."
            },
            "required": false,
            "name": "desplazamiento",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Una página de accesos, del más reciente al más antiguo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaAcceso"
                }
              }
            }
          },
          "400": {
            "description": "Petición inválida. El cuerpo o los parámetros no pasan la validación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "401": {
            "description": "Falta la sesión o el token no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "403": {
            "description": "La sesión es válida pero el rol no alcanza, o la cuenta está inactiva.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "503": {
            "description": "El servicio o la base de datos no están disponibles.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/usuarios": {
      "get": {
        "tags": [
          "Administrador Maestro"
        ],
        "summary": "Listado paginado de usuarios, con filtros",
        "description": "El filtro y el recorte se aplican en la base de datos. `total` es el número de filas que cumplen el filtro, no las devueltas: sin él la pantalla no puede saber cuántas páginas quedan. La búsqueda es un `ilike` sin índices trigram; suficiente para un centro de formación y anotado por si la tabla crece.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "docente",
                "estudiante"
              ],
              "description": "Filtra por rol. Sin este parámetro, todos los roles."
            },
            "required": false,
            "name": "rol",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "Filtra por estado de la cuenta. Un valor distinto de `true` o `false` se rechaza con 400."
            },
            "required": false,
            "name": "activo",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Texto libre sobre nombres, apellidos, correo y cédula. Insensible a mayúsculas."
            },
            "required": false,
            "name": "busqueda",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "description": "Tamaño de página. Por defecto 25."
            },
            "required": false,
            "name": "limite",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0,
              "description": "Filas a saltar antes de esta página. Por defecto 0."
            },
            "required": false,
            "name": "desplazamiento",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Una página de usuarios, ordenados por apellido y nombre.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaUsuarios"
                }
              }
            }
          },
          "400": {
            "description": "Petición inválida. El cuerpo o los parámetros no pasan la validación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "401": {
            "description": "Falta la sesión o el token no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "403": {
            "description": "La sesión es válida pero el rol no alcanza, o la cuenta está inactiva.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "503": {
            "description": "El servicio o la base de datos no están disponibles.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/usuarios/invitaciones": {
      "post": {
        "tags": [
          "Administrador Maestro"
        ],
        "summary": "Invita a un docente por correo (token de un solo uso)",
        "description": "El administrador aporta el correo. El servidor genera un token, guarda SOLO su huella en la base y envía el enlace de activación. El enlace también se devuelve en la respuesta para poder probar el flujo aunque el correo no llegue.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CuerpoCorreoInvitacion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invitación creada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaInvitacion"
                }
              }
            }
          },
          "400": {
            "description": "Petición inválida. El cuerpo o los parámetros no pasan la validación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "401": {
            "description": "Falta la sesión o el token no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "403": {
            "description": "La sesión es válida pero el rol no alcanza, o la cuenta está inactiva.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          },
          "503": {
            "description": "El servicio o la base de datos no están disponibles.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuerpoError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}